+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-3074] Endpoint to share notifications
+
 ### Changed
 
 - [TD-3163] Auth tokens now include `role` claim instead of `is_admin` flag